2020/08/28 周五

+++ Parallel_Programming_for_FPGAs/Chapter2_FIR/experiment2_SimpleAdd/head.h
@@ -2,4 +2,6 @@
 // typedef short T;
 // typedef int T;
 // typedef long T;
-typedef long long T;
+// typedef long long T;
+// typedef float T;
+typedef double T;
